Update Number of appearances of digit 2.py
Count the number of appearances of digit 2 in all numbers that are between 0 and n. It could be improved in this way: we can define the number of the appearances of digit 2 in all numbers between 0 and 25, then between 0 and 50, 0 and 100 and 0 and 1000. Now, we can easily calculate for bigger numbers like: 100000100, because we can split it in: 100 * 1000 * 1000 + 100. But we know how many 2-digits are in these intervals. :)